moving char_buffer to utils.
