+++ cclient/response/BUILD
@@ -9,5 +9,6 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//cclient/types",
+        "//utils:char_buffer",
     ],
 )
