+++ cclient/response/error.h
@@ -18,6 +18,7 @@
 #define CCLIENT_RESPONSE_ERROR_H
 
 #include "cclient/types/types.h"
+#include "utils/char_buffer.h"
 
 #ifdef __cplusplus
 extern "C" {
