+++ cclient/service.c
@@ -11,8 +11,6 @@
 #include "utils/logger_helper.h"
 
 retcode_t iota_client_service_init(iota_client_service_t* const serv) {
-  // enable logger
-  logger_init_types();
   // init serializer
   if (serv->serializer_type == SR_JSON) {
     logger_init_json_serializer();
@@ -28,5 +26,4 @@ void iota_client_service_destroy(iota_client_service_t* const serv) {
   if (serv->serializer_type == SR_JSON) {
     logger_destroy_json_serializer();
   }
-  logger_destroy_types();
 }
